[driver] WaitFor(future)
Support waiting for a std::future in tests, alongside waiting for sync
completions.

The main purpose is to wait on a std::future returned by a
DispatcherBound::AsyncCall, while also running the dispatchers in the
driver runtime, if any.

Rewrote one test to exercise the WaitFor(future). Looking for
suggestions as to where to put more focused unit tests.

Change-Id: I58cd5fef681381af4d5a2156b22402c1f95ec5c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811848
Reviewed-by: Novin Changizi <novinc@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>