[kbma] Introduce counter increment/decrement logic.
This CL introduces the hooks through which attribution objects
are incremented and decremented. The instrumentation through
which cow pages are given an attribution object will be introduced
in a follow up cl, so for now even with kbma enabled this cl
would be a no-op.

Bug: 117196

Change-Id: I1f2d9545cd1ac8ca62ab7e1e5bf040187e283338
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802267
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Luke Nicholson <lukenicholson@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>