[time] Register the pull time source
The pull time source is registered as a new component
in the product assembly, and the core size limits
calculations.

I expect little effect on the size of the time bundle,
as the component is identical to the push component,
and the config files should be small.

Bug: 113956
Change-Id: I2608116cd8239c1e97497c791c3b9ff183cbefe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803368
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Size-Review: Aaron Wood <aaronwood@google.com>