[botanist] Get QEMU kernel using image name and CPU
The kernel that botanist passed to QEMU was previously just the first
image with type `qemu-kernel` in images.json. This is a problem when
multiple QEMU kernels are present with different target architectures
(as in the case of our RISC-V builder). This CL solves the problem by
filtering by both the image name and the CPU of the associated image.

Bug: b/270961878
Change-Id: Ic7c26d18a87a186e649ac9ee693580df3f8a013e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812849
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>