[netstack3] Get Buffer len and capacity atomically
Without this, a future change to make the RingBuffer held in a
SendBufferWithZirconSocket resizable would introduce raciness. Since
both len() and capacity() would depend on the fullness of the zircon
socket, and that can change due to the process attached to the other end
reading, the values returned by len() and capacity() would not be
guaranteed to be consistent with one another. Prevent this by, in this
change, returning both in the result of a single method.

Bug: 115456
Change-Id: Ie5b4b4ed6bdc05dd9d8c9d338e5d52ac323173f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812987
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>