[botanist] Have CSO builder using CSO-only mode.
The CSO builder is using ffx experiment level 4, and since the rest of
the ffx builders are running with CSO in enabled mode, we can use the
CSO builder to test the CSO-only mode.
Also move enabled mode to experiment level 1 so we can see if it helps
with flashing.

Change-Id: Ib51e2ea33fdc21713bf71859f80facc708c2442a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813054
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>