[msd-arm-mali] Process job interrupts using port
Job interrupts were waited on by the job interrupt thread, which posted
a request to the device thread and waited for a reply. The extra work
takes at least 25 us of CPU time per interrupt (on astro), which adds up
over the course of 10-20 interrupts for frame.

Instead we should bind the job interrupt to a port, so the device thread
can wait on it directly.

Bug: 122629
Change-Id: I6d76bc703699e62e84626deec4d45b43d25ca657
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811809
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>