[gn_desc] Optimizations to increase startup speed by 60 percent
Changes the optimization level to 'speed' for the binary and it's rlibs
as well as reading the entire json file into a String buffer before it
starts to parse it with serde.

Half the improvement is from reading into a String, the other half from
the increased optimization level.

Change-Id: I87aac7ae78c374ea7653a38a0c8ce496780aa7f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813368
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>