[uart] Introduce SyncPolicy
A SyncPolicy provides the mechanisms required for
synchronization. That is:
    * Lock Type: Opaque capability
    * Guard Type: Opaque RAII for acquiring the lock
    * Waiter Type: Mechanism for waiting on
                   Tx Hw Fifo becoming non full.
    * AssertHeld: Assert a Lock instance is held.
    * DefaultLockPolicy: Default Lock Policy for methods.

The Guard type is allowed to have one template argument, which is called
LockPolicy. A LockPolicy is defined on the driver methods, such that
the same object can use different flavors. This allows for the
kernel environment to implement:
    * IrqSave
    * NoIrqSave
    * NoopLocking (Panic Shell)
See following patch for details.

Bug: 89182
Test: fx test uart-unittests
Change-Id: Iccc0db8e0cad7be594a3b2b33ed12bf54a0e8ccd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803047
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>