[magma] Move C API notification sending to the calling thread.
Writing the the channel is thread-safe, so we don't have to post it to
the connection thread. This should at least 16 us of work per
notification on Nelson.

This patch is a port of I0e7df0a1141ad45cfc55785176c73d26bd5f4a0f to
the MSD C API, so it's easier to backport to release branches.

Bug: 122629
Change-Id: I4941aafef22cd9aadf3341a83ab55d4cfe283cfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813601
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>