[benchmarks] Implement dumpFile() using SSH instead of SL4F
Add an implementation of dumpFile() that uses SSH instead of the SL4F
server, and use it for the common cases.

This will be useful for avoiding the dependency on SL4F for the
simpler performance tests that just run a component and don't use
tracing.  For now, those simpler tests still start the SL4F server,
but I will address that in a later CL.

Bug: 122165
Change-Id: I7b674f267e860e390993b9fbfb1def0476d81549
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813447
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>