[wlan][wlancfg] Fix counter stats all 0's bug
When UpdateExperiment is called, new instances of Cobalt metric logger
and stats are created, but the LazyNodes in Inspect are not updated to
point to the new instance of the stats. In this commit, we only refresh
the Cobalt logger and not the stats so that the bug no longer occurs.

However, this only fits with the current usage pattern where
UpdateExperiment is only called at the beginning. It's possible that in
the future it'll be called long after device bootup, at which point this
implementation is no longer sufficient because non-experiment and
experiment stats are intermingled. See fxbug.dev/122985 for discussion
on handling this long term.

This commit also includes a new unit test, and the Inspect LazyNodes
with the counters are also now kept in StatsLogger, to reduce the chance
of introducing the same bug again in the future.

Bug: 120678
Test:
1. Added unit test
2. Verified bug is fixed on nelson

Change-Id: I0b1aeccdb4e17e96f25e5e06e46d1040d2b33afc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814212
Commit-Queue: Kiet Tran <kiettran@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>