[virtio-block] Fix invoking block impl callback twice.
In the corner case of performing a block impl IO of 0 blocks, the block
impl callback is invoked twice; once with an error code and the next
with a success code (ZX_OK).

Fix this while using block::CheckIoRange() to check the validity of the
IO range.

Bug: 43065
Test: fx test virtio-block-test
Change-Id: Ie3035471d414fb53c0593d55e3eafa93b179dc71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814454
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>