[memory_sampler] Add a service to collect memory profile samples
This commit exposes a new FIDL interface for a process to report its
memory allocations, as well as a component which serves this FIDL
interface.

The component collects allocation information from its client's
process and, once the client stops, creates a `pprof`-compatible
protobuf profile that it persists to a file.

This is expected to be the first change of a series; the component is
currently not routed nor used anywhere, but would eventually expose a
generic mechanism for arbitrary components to opt into profiling.

Test: fx test memory_sampler
Change-Id: I04407a45304263d18417b9ea6786ff063d9fba00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801327
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>