[botanist] Split Target.TestConfig into BaseTarget
Break apart TestConfig from the Target interface into a smaller
interface that allows for a partial implementation. The resulting
slimmed-down interface is embedded into Target for backwards
compatibility with current uses of Target.

This is a necessary step to support generic targets in an upcoming CL.
Generic targets are not guaranteed to be running Fuchsia, but botanist
still needs to export information about them to the testbed config for
tests to injest.

Rename variables to make the type of target clear. There will future
changes to reduce naming redundancy.

Change-Id: Ic6ad8fe3b5645b3485e8e5af1ec07d282c4e079b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813948
Commit-Queue: Sam Balana <sbalana@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>