[fx][pkg] Check if the server is running
We now have a way to run the ffx repository server on arbitrary ports,
as well as a way to check if the repository server is running. This
changes `fx serve-updates`, `fx add-update-source`, and `fx-serve-remote`
to use these tools to allow the port to be configured, and to only error
out if we actually have a port conflict, rather than a potential conflict
in in between a local and remote repository server config.

Bug: 82788
Bug: 119089
Change-Id: I813c8d8f332eff1a9d17521d71ad82520ca1b7a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791846
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>