[build] Change build/cat.sh to atomically write files
This changes the build to atomically write the all_package_manifests.list.
This is primarily useful for the incremental publisher, which watches
this file for changes. Since this file format is line oriented, there
isn't a way to know if the file was truncated, or fully written. This
means that if the incremental publisher manages to read the file before
it's fully written, it will only try to publish a subset of changes. By
atomically writing this file, we should be protected from this state.

Long term though we should change the file format so we can tell if
this file was truncated to catch this error with out-of-tree package
manifest lists.

Fixed: 123151
Change-Id: I4ded64a96c7888f91da5716dc3234adc147db4e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815272
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>