[botanist] Remove target naming redundancies
Remove trailing "Target" from type names within
//tools/botanist/targets. This information is already captured by the
name of the package. For example, `targets.QEMU` conveys the same
information as `targets.QEMUTarget`.

Rename target interfaces for the same reason as above. In this case, we
can be more specific with the naming; "Target" is renamed to "Fuchsia"
to convey constraints noted in the comments.

Change-Id: I1f0d6b96a1f8598e0835671d12ea5a718bdd766d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814007
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Sam Balana <sbalana@google.com>