[kernel][vm] Split interval and add page
The new LookupOrAllocateCheckForInterval function can add pages in the
middle of an interval by splitting the interval around the offset that
is being looked up. If the offset does not fall in an interval, calling
this function is similar to calling LookupOrAllocate.

Not used outside of unit tests.

Test: vmpl_unittest
Bug: 122842
Run-All-Tests: true
Change-Id: I05172067c9ca308e29f87a57d5ce8e89b78eba26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812808
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>