[media/ffmpeg] restore required 'decoders'
Little-endian PCM decoders were dropped from the max and default
ffmpeg profiles, because we don't need those decoders. Those formats
can be consumed directly by our audio renderer. As it turns out, the
demux will not identify LE PCM streams correctly if these decoders
are not enabled. This breaks a test, so those decoders are restored
in this CL.

When we create product-specific profiles for smart displays, we won't
need to demux these formats, so we'll be able to drop those formats,
as well as most others.

Currently, the ffmpeg roller is broken, because we are using a config
that lacks these decoders. Once this CL lands, another roll will be
done, and the ffmpeg roller should succeed.

Test: build issue, manually tested
Change-Id: Iaf96d13d2ee785a01d4107a852cf79027753da84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816074
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>