[starnix] Disable flaky file locking tests and add reduced testcase
The fcntl file locking tests are flaky in Starnix right now because we
do not release file locks on process termination consistently. This adds
a reduced testcase for the specific issue and marks the tests "skip" in
the expectations file to turn the bots green while we investigate a
solution.

Bug: 122600
Change-Id: I85221ed17ae4ce309694a9784a0b0820e58fbd4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816568
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>