[drivers] Add an allowlist on platform device banjo
Add an allowlist to the banjo protocol. This makes it more obvious which
devices still need to be migrated.

Remove the vast majority of direct dependencies on this banjo
protocol, because these drivers depend on device-protocol-pdev.
Some of these drivers still need to be migrated to FIDL as well.

Bug: 110796
Change-Id: I71c7b64a86bc14b633053874875134c8abdf88cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816887
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>