[assembly][recovery] Switch recovery images to utility feature-set-level
This should be a no-op, this switches the recovery-like images over
to using the 'utility' feature_set_level (which is currently the
same as bootstrap).  After this, and the equivalent change for OOT
images, the packages can be moved from the transitional labels
over to Assembly.

Bug: 122861
Change-Id: Ib1a9e601ab3c40e131321eb92704ea8c7e943de1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816588
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>