[media] Define and use scheduler roles for mixers and dispatch.
This change is an incremental step towards adopting the scheduler role
API in audio core. Two roles are defined for the mixer and dispatch
threads, respectively, with the currently hard-coded default scheduling
parameters added to core.profiles.

Although this change obviates the "mix_profile" option in the audio core
process config, that code is left in place for potential reuse as inputs
to the future parameterized scheduler role API. What form this takes is
left to a future change by the audio core owners.

A separate change in the product profile configs will address product
performance requirements in the referenced bug.

Bug: b/268386463
Change-Id: I0044f9418706a22fdfae9a5a2494acbae5dbf7dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/808594
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>