[starnix] Work on ARM64 compilation.
Remove definition of MINSIGSTKSZ from signals/syscalls.rs. This is
defined in the generated uapi files so we can avoid maintaining our own
definition. The disadvantage is that the type is u32 so we now need to
cast to usize when we use it.

Correct the definition of nlink_t and use it (previously the x64 didn't
use this type because it was wrong).

Use blksize_t definition where appropriate.

Change-Id: Ie9537f7b868a68b8ad6da5b3d6f869c7cb4c1f8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817027
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>