[a11y] Adds Gesture Handler v2
This change is part of the migration of a11y services to the new gesture disambiguation protocols.

This change is just a copy of the old Gesturehandler, changing a few data types to match the new expected format (mostly GestureContext).

Once the migration is done, this will be renamed to be back the Gesturehandler, and exists as it is for now so that we can support both gd protocols for a while.

Bug: 100030
Change-Id: I8f859ba9d2ee04bee019cdb381c2d653966b8c60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801596
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>