[netstack] Add config-data for GOMAXPROCS
Add a config-data directory for netstack and use it to configure
GOMAXPROCS, which is the number of operating system threads that can be
used to run Go worker threads simultaneously. Expose the field in
inspect, as well as the number reported by zx_system_get_num_cpus.

For now, providing config-data to the netstack is optional. This is
because there exist many usages of a hermetic netstack, such as in
out-of-tree tests, that do not currently provide config-data and which
would need to provide test-specific data if it was required. Instead of
making it required, default to not modifying GOMAXPROCS. Also, make the
GOMAXPROCS field of the config-data itself optional, such that in the
default scenario, an empty configuration is provided and the netstack
uses the Go runtime's default value.

Bug: 123194
Test: netstack-inspect-integration-test
Change-Id: I1bd81fe01c3b7b5154a2cbe24ce0e2a28c05ec91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815847
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Size-Review: David Iserovich <diserovich@google.com>
Size-Review: Aaron Wood <aaronwood@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>